Improve test module structure
