+++ tests/block_tests/mod.rs
@@ -1,0 +1,1 @@
+mod block_parser;
