+++ tests/mod.rs
@@ -1,2 +1,2 @@
-mod block_parser;
-mod state;
+mod block_tests;
+mod state_tests;
