+++ tests/state_tests/root_tree_extensions/mod.rs
@@ -1,0 +1,2 @@
+mod simple_improper_extension;
+mod simple_proper_extension;
