+++ tests/state_tests/root_tree_extensions/simple_improper_extension.rs
@@ -1,0 +1,12 @@
+// TODO simple_improper_extension
+#[tokio::test]
+async fn test() {
+    // ----- Dangling Branches -----
+    //   Before  |   After
+    // ------- indices ------
+    //     0     |     0
+    // ----------------------
+    //     0     =>    0
+    //     |     =>   / \
+    //     1     =>  1   2
+}
